Simplify Search Labels action
