+++ front/types/assistant/conversation.ts
@@ -3,10 +3,10 @@ import type { DustAppRunActionType } from "@app/lib/actions/dust_app_run";
 import type { ProcessActionType } from "@app/lib/actions/process";
 import type { ReasoningActionType } from "@app/lib/actions/reasoning";
 import type { RetrievalActionType } from "@app/lib/actions/retrieval";
+import type { SearchLabelsActionType } from "@app/lib/actions/search_labels";
 import type { TablesQueryActionType } from "@app/lib/actions/tables_query";
 import type { ConversationIncludeFileActionType } from "@app/lib/actions/types/conversation/include_file";
 import type { ConversationListFilesActionType } from "@app/lib/actions/types/conversation/list_files";
-import type { SearchLabelsActionType } from "@app/lib/actions/types/search_labels";
 import type { WebsearchActionType } from "@app/lib/actions/types/websearch";
 
 import type { ContentFragmentType } from "../content_fragment";
